backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
asyncpg
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt
python-dotenv
email-validator
orjson
